example/requirements.txt
Flask==0.10.1
Flask-Assets==0.9
Flask-SQLAlchemy==2.0
graphql-core==0.4.12.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
six==1.10.0
SQLAlchemy==1.0.6
webassets==0.9
Werkzeug==0.11.4
wheel==0.24.0
